Move parse_nfs_url to the new core/payload file
The core/util.py file is overloaded so move payload specific utilities
to core/payload. These will be specific for payload but usable by module
and main process too.

Move tests for this from util_test to payload_test.
