+++ pyanaconda/payload/dnf/payload.py
@@ -53,6 +53,7 @@ from pyanaconda.core.configuration.anaconda import conf
 from pyanaconda.core.constants import INSTALL_TREE, ISO_DIR, DRACUT_REPODIR, DRACUT_ISODIR, \
     PAYLOAD_TYPE_DNF
 from pyanaconda.core.i18n import N_, _
+from pyanaconda.core.payload import parse_nfs_url
 from pyanaconda.core.regexes import VERSION_DIGITS
 from pyanaconda.core.util import ProxyString, ProxyStringError, decode_bytes
 from pyanaconda.flags import flags
@@ -1595,7 +1596,7 @@ class DNFPayload(Payload):
         path = None
 
         if iso_device_path and repo_device_path:
-            path = util.parse_nfs_url('nfs:%s' % iso_device_path)[2]
+            path = parse_nfs_url('nfs:%s' % iso_device_path)[2]
             # See if the dir holding the iso is what we want
             # and also if we have an iso mounted to /run/install/repo
             if path and path in iso_device_path and DRACUT_ISODIR in repo_device_path:
@@ -1605,7 +1606,7 @@ class DNFPayload(Payload):
             # see if the nfs dir is mounted
             need_mount = True
             if repo_device_path:
-                _options, host, path = util.parse_nfs_url('nfs:%s' % repo_device_path)
+                _options, host, path = parse_nfs_url('nfs:%s' % repo_device_path)
                 if method.server and method.server == host and \
                    method.dir and method.dir == path:
                     need_mount = False
@@ -1615,7 +1616,7 @@ class DNFPayload(Payload):
                 # nfs mount have changed. It is already mounted, but on INSTALL_TREE
                 # which is the same as DRACUT_ISODIR, making it hard for _setup_NFS
                 # to detect that it is already mounted.
-                _options, host, path = util.parse_nfs_url('nfs:%s' % iso_device_path)
+                _options, host, path = parse_nfs_url('nfs:%s' % iso_device_path)
                 if path and path in iso_device_path:
                     need_mount = False
                     path = DRACUT_ISODIR
@@ -1715,7 +1716,7 @@ class DNFPayload(Payload):
                     # prepend nfs: to the url as that's what the parser
                     # wants.  Note we don't get options from this, but
                     # that's OK for the UI at least.
-                    _options, host, path = util.parse_nfs_url("nfs:%s" % repo_device_path)
+                    _options, host, path = parse_nfs_url("nfs:%s" % repo_device_path)
                     method.method = "nfs"
                     method.server = host
                     method.dir = path
