+++ pyanaconda/payload/source/factory.py
@@ -16,7 +16,7 @@
 # License and may only be used or replicated with the express permission of
 # Red Hat, Inc.
 #
-from pyanaconda.core.util import parse_nfs_url
+from pyanaconda.core.payload import parse_nfs_url
 from pyanaconda.payload.source.sources import CDRomSource, HDDSource, NFSSource, HTTPSource, \
     HTTPSSource, FTPSource, FileSource, HMCSource
